Bug#37201818 Can not create table with 256 or more blob columns
Creating table with 256 or more blob columns overflow NdbTableImpl::m_noOfBlobs.
Need to reduce blob inline size with use COMMENT 'NDB_COLUMN=BLOB_INLINE_SIZE=100'.
Causing table creation to fail with:

  Error 1005 Can't create table 't1' (use SHOW WARNINGS for more info).
  Warning 1296 Got error 4263 'Invalid blob attributes or invalid blob parts table' from NDB
  Warning 1296 Failed to rollback after CREATE TABLE failure.

+++ storage/ndb/src/ndbapi/NdbDictionaryImpl.hpp
@@ -314,7 +314,7 @@ class NdbTableImpl : public NdbDictionary::Table, public NdbDictObjectImpl {
   Uint8 m_noOfKeys;
   // if all pk = dk then this is zero!
   Uint8 m_noOfDistributionKeys;
-  Uint8 m_noOfBlobs;
+  Uint16 m_noOfBlobs;
   Uint16 m_noOfDiskColumns;
   Uint8 m_replicaCount;
